pyproject.toml
[project]
name = "pyconverter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "ffmpeg-python>=0.2.0",
]
